api/requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.32.1
sqlalchemy==2.0.37
alembic==1.14.0
psycopg2-binary==2.9.10
redis==5.2.1
rq==1.16.2
jinja2==3.1.4
python-multipart==0.0.20
passlib==1.7.4
pydantic-settings==2.7.0
openai>=1.108.0,<2.0.0
email-validator==2.2.0
httpx==0.27.2
PyJWT[crypto]==2.10.1
itsdangerous==2.2.0
pytest==8.3.4
cryptography==44.0.0
PyYAML==6.0.2
